chore(linear): move mcp-client + tool-names into issue/ (rename only)
ADR-0003 step 3. The Linear MCP client wrapper and the tool-name
constants now live in src/issue/. The src/linear/ copies are
single-line re-export shims so existing import paths keep resolving
unchanged.

The mcp-client test moved with the source (it doesn't import from
any path that needs updating yet — that's step 4). sync-pull,
sync-push, and integration tests stay in src/linear/ for now;
they'll be rewritten in step 4 to use Issue.Service.
